Fixup a few problems with EAP state transition
First of all we don't need to satisfy valid EAP state flags for
IKEV2_STATE_EAP as it's an initial EAP exchange state.  Then
when waiting for the "ca" process to construct our AUTH payload
we need to bail while sa_localauth is not available.  With this
change Win7 is able to establish the the tunnel again.

ok markus

Imported from: https://github.com/openbsd/src/commit/c26c395bce59