iked: remove another unused variable
It is not entirely obvious if some logic was forgotten or if that variable
was added for debug logging or... Be that as it may, as tobhe points out,
this variable has been unused since the initial import of iked. The logic
seems sound without it, so it has never served a purpose and can go.

ok tobhe

Imported from: https://github.com/openbsd/src/commit/2eafc8b460af