Minimal fix to make internal hyperlinking work for these sections: Macros, Global configuration, Automatic keying policies. To this end, switch .Sh to .Ss and .Sy to .Sx for these sections. While here, also fix the internal link .Sx RADIUS CONFIGURATION .
There are many custom sections that are not mentioned in the
table of contents and that remain on the .Sh level.  I'm not
touching that because i'm unsure whether that is intentional
or accidental.  Maybe an iked(8) expert should look into that.

OK florian@

Imported from: https://github.com/openbsd/src/commit/e6f322b84492